Paper/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ctw-va"
version = "0.1.0"
description = "Civatas-TW Vendor Audit — standalone experimental platform for 5-vendor LLM simulation of Taiwan 2024 presidential election"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "CH Tseng", email = "chtseng.neural@gmail.com" }]
dependencies = [
    "click>=8.1",
    "openai>=1.40",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "pandas>=2.1",
    "numpy>=1.26",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "pydantic>=2.5",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
]

[project.scripts]
civatas-exp = "ctw_va.cli.__main__:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
